# openapi/connector_spec.yaml
openapi: 3.0.1
info:
  title: Connector API
  version: 1.0.0
paths:
...
  /v1/api/accounts:
    get:
      operationId: searchGatewayAccounts
      parameters:
      - in: query
        name: accountIds
        schema:
          type: string
          description: Comma separate list of gateway account IDs
          example: "1,2"
          pattern: "^[\\d,]+$"
      - in: query
        name: serviceIds
        schema:
          type: string
          description: Comma separated list of service external IDs
          example: "46eb1b601348499196c99de90482ee68,service-external-id-2"
          pattern: "^(?:[A-z0-9]+,?)+$"
      - in: query
        name: moto_enabled
        schema:
          type: string
          description: The accounts will be filtered by whether or not MOTO payment
            are enabled for the account if this parameter is provided. "true" or "false"
          example: "true"
          pattern: true|false
      - in: query
        name: apple_pay_enabled
        schema:
          type: string
          description: The accounts will be filtered by whether or not Apple pay is
            enabled for the account if this parameter is provided. "true" or "false".
          example: "true"
          pattern: true|false
      - in: query
        name: google_pay_enabled
        schema:
          type: string
          description: The accounts will be filtered by whether or not Google pay
            is enabled for the account if this parameter is provided. "true" or "false".
          example: "true"
          pattern: true|false
      - in: query
        name: requires_3ds
        schema:
          type: string
          description: The accounts will be filtered by whether or not 3DS is required
            for the account if this parameter is provided. "true" or "false".
          example: "true"
          pattern: true|false
      - in: query
        name: type
        schema:
          type: string
          description: The accounts will be filtered by type if this parameter is
            provided. "test" or "live".
          example: live
          pattern: live|test
      - in: query
        name: payment_provider
        schema:
          type: string
          description: "The accounts will be filtered by payment provider if this\
            \ parameter is provided. One of \"sandbox\" or \"worldpay\", \"smartpay\"\
            , \"epdq\" or \"stripe\""
          example: live
          pattern: sandbox|worldpay|smartpay|epdq|stripe
      - in: query
        name: payment_provider_account_id
        schema:
          type: string
          description: Accounts will be filtered by payment provider account ID
          example: payment-provider-account-id
      - in: query
        name: provider_switch_enabled
        schema:
          type: string
          description: The accounts will be filtered by whether or not payment provider
            switch is enabled for the account if this parameter is provided. "true"
            or "false".
          example: "true"
          pattern: true|false
      - in: query
        name: recurring_enabled
        schema:
          type: string
          description: The accounts will be filtered by whether or not recurring payments
            are enabled for the account when this parameter is provided.
          example: "true"
          pattern: true|false
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayAccountsListDTO"
          description: OK
      summary: Search gateway accounts
      tags:
      - Gateway accounts
    post:
      operationId: createNewGatewayAccount
      parameters:
      - in: query
        name: degatewayification
        schema:
          type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GatewayAccountRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateGatewayAccountResponse"
          description: Created
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing required fields or invalid values
      summary: 'Create a new gateway account '
      tags:
      - Gateway accounts
  /v1/api/accounts/{accountId}:
    get:
      description: "Get gateway account by internal ID. Returns notifications credentials,\
        \ gateway account credentials (without password). Doesn't include card_types\
        \ or gateway_merchant_id"
      operationId: getGatewayAccount
      parameters:
      - description: Gateway account ID
        example: 1
        in: path
        name: accountId
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayAccountWithCredentialsWithInternalIdResponse"
          description: OK
        "404":
          description: Not found
      summary: Find gateway account by ID
      tags:
      - Gateway accounts
    patch:
      description: "A generic endpoint that allows the patching of allow_apple_pay,\
        \ allow_google_pay, block_prepaid_cards, notify_settings, email_collection_mode,\
        \ corporate_credit_card_surcharge_amount, corporate_debit_card_surcharge_amount,\
        \ corporate_prepaid_debit_card_surcharge_amount, allow_zero_amount, allow_moto,\
        \ moto_mask_card_number_input, moto_mask_card_security_code_input, allow_telephone_payment_notifications,\
        \ send_payer_ip_address_to_gateway, send_payer_email_to_gateway, integration_version_3ds,\
        \ send_reference_to_gateway, allow_authorisation_api or worldpay_exemption_engine_enabled\
        \ using a JSON Patch-esque message body."
      operationId: patchGatewayAccountByGatewayAccountId
      parameters:
      - description: Gateway account ID
        example: 1
        in: path
        name: accountId
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: string
              example:
                op: replace
                path: allow_apple_pay
                value: true
      responses:
        "200":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request
        "404":
          description: Not found
      summary: 'Patch a gateway account '
      tags:
      - Gateway accounts
...
  /v1/api/service/{serviceId}/account/{accountType}:
    get:
      description: "Get gateway account by service external ID and account type (test|live).\
        \ Returns notifications credentials, gateway account credentials (without\
        \ password). Doesn't include card_types or gateway_merchant_id"
      operationId: getGatewayAccountByServiceIdAndAccountType
      parameters:
      - description: Service ID
        example: 46eb1b601348499196c99de90482ee68
        in: path
        name: serviceId
        required: true
        schema:
          type: string
      - description: Account type
        example: test
        in: path
        name: accountType
        required: true
        schema:
          type: string
          enum:
          - test
          - live
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatewayAccountWithCredentialsResponse"
          description: OK
        "404":
          description: Not found
      summary: Find gateway account by service external ID and account type (test|live)
      tags:
      - Gateway accounts
    patch:
      description: "A generic endpoint that allows the patching of allow_apple_pay,\
        \ allow_google_pay, block_prepaid_cards, notify_settings, email_collection_mode,\
        \ corporate_credit_card_surcharge_amount, corporate_debit_card_surcharge_amount,\
        \ corporate_prepaid_debit_card_surcharge_amount, allow_zero_amount, allow_moto,\
        \ moto_mask_card_number_input, moto_mask_card_security_code_input, allow_telephone_payment_notifications,\
        \ send_payer_ip_address_to_gateway, send_payer_email_to_gateway, integration_version_3ds,\
        \ send_reference_to_gateway, allow_authorisation_api, worldpay_corporate_exemptions_enabled\
        \ or worldpay_exemption_engine_enabled using a JSON Patch-esque message body."
      operationId: patchGatewayAccountByServiceIdAndType
      parameters:
      - description: Service ID
        example: 46eb1b601348499196c99de90482ee68
        in: path
        name: serviceId
        required: true
        schema:
          type: string
      - description: Account type
        example: test
        in: path
        name: accountType
        required: true
        schema:
          type: string
          enum:
          - test
          - live
      requestBody:
        content:
          application/json:
            schema:
              type: string
              example:
                op: replace
                path: allow_apple_pay
                value: true
      responses:
        "200":
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad request
        "404":
          description: Not found
      summary: 'Patch a gateway account '
      tags:
      - Gateway accounts
...
  /v1/api/service/{serviceId}/account/{accountType}/adyen-setup/{credentialExternalId}:
    get:
      operationId: getAdyenAccountSetup
      parameters:
      - description: Service ID
        example: 46eb1b601348499196c99de90482ee68
        in: path
        name: serviceId
        required: true
        schema:
          type: string
      - description: Account type
        example: test
        in: path
        name: accountType
        required: true
        schema:
          type: string
          enum:
          - test
          - live
      - description: Credential External ID
        example: 46eb1b601348499196c99de90482ee68
        in: path
        name: credentialExternalId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              example:
                service_id: 46eb1b601348499196c99de90482ee68
                credential_external_id: 46eb1b601348499196c99de90482ee68
                gateway_account_id: 123
                tasks:
                  organisation_details:
                    status: COMPLETED
                  legal_terms:
                    status: COMPLETED
                  bank_details:
                    status: COMPLETED
                  responsible_person:
                    status: COMPLETED
                  director:
                    status: COMPLETED
                  reason_for_taking_payments:
                    status: COMPLETED
              schema:
                $ref: "#/components/schemas/AdyenAccountSetupResponse"
          description: OK
        "404":
          description: Not found
      summary: "Retrieve Adyen account setup tasks for a given service ID, account\
        \ type and credential ID"
      tags:
      - Gateway accounts
    patch:
      description: "Support patching following paths: <br>bank_account, responsible_person,\
        \ vat_number, company_number, director, government_entity_document, organisation_details"
      operationId: patchAdyenAccountSetup
      parameters:
      - description: Service ID
        example: 46eb1b601348499196c99de90482ee68
        in: path
        name: serviceId
        required: true
        schema:
          type: string
      - description: Account type
        example: test
        in: path
        name: accountType
        required: true
        schema:
          type: string
          enum:
          - test
          - live
      - description: Credential External ID
        example: 46eb1b601348499196c99de90482ee68
        in: path
        name: credentialExternalId
        required: true
        schema:
          type: string
      requestBody:
        content:
          '*/*':
            example:
            - op: replace
              path: bank_details
              value: COMPLETED
            - op: replace
              path: responsible_person
              value: NOT_STARTED
            schema:
              $ref: "#/components/schemas/AdyenSetupPatchRequest"
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not found
      summary: "Update Adyen account setup tasks for a given service ID, account type\
        \ and credential ID"
      tags:
      - Gateway accounts
...
components:
  schemas:
...
    CreateGatewayAccountResponse:
      type: object
      properties:
        analytics_id:
          type: string
          example: ananytics-id
        description:
          type: string
          example: account for some gov org
        external_id:
          type: string
          example: ab2c296ed98647e9a25f045f5e6e87a2
        gateway_account_id:
          type: string
          example: "2"
        links:
          type: array
          example:
          - href: https://connector.url/v1/api/accounts/2
            method: GET
            rel: self
          items:
            type: object
            additionalProperties:
              type: object
              example:
              - href: https://connector.url/v1/api/accounts/2
                method: GET
                rel: self
            example:
            - href: https://connector.url/v1/api/accounts/2
              method: GET
              rel: self
        requires_3ds:
          type: boolean
          example: true
        send_payer_email_to_gateway:
          type: boolean
          example: true
        send_payer_ip_address_to_gateway:
          type: boolean
          example: true
        service_name:
          type: string
          example: service name
        type:
          type: string
          example: live
...
    ErrorResponse:
      type: object
      properties:
        error_identifier:
          type: string
          enum:
          - ACCOUNT_DISABLED
          - ACCOUNT_NOT_LINKED_WITH_PSP
          - AGREEMENT_NOT_ACTIVE
          - AGREEMENT_NOT_FOUND
          - AMOUNT_BELOW_MINIMUM
          - AUTH_TOKEN_INVALID
          - AUTH_TOKEN_REVOKED
          - AUTHORISATION_API_NOT_ALLOWED
          - AUTHORISATION_ERROR
          - AUTHORISATION_REJECTED
          - AUTHORISATION_TIMEOUT
          - CANCEL_CHARGE_FAILURE_DUE_TO_CONFLICTING_TERMINAL_STATE_AT_GATEWAY_CHARGE_STATE_FORCIBLY_TRANSITIONED
          - CANCEL_CHARGE_FAILURE_DUE_TO_CONFLICTING_TERMINAL_STATE_AT_GATEWAY_INVALID_STATE_TRANSITION
          - CARD_NUMBER_IN_PAYMENT_LINK_REFERENCE_REJECTED
          - CARD_NUMBER_REJECTED
          - GENERIC
          - IDEMPOTENCY_KEY_USED
          - INCORRECT_AUTHORISATION_MODE_FOR_SAVE_PAYMENT_INSTRUMENT_TO_AGREEMENT
          - INVALID_ATTRIBUTE_VALUE
          - MISSING_MANDATORY_ATTRIBUTE
          - MOTO_NOT_ALLOWED
          - NON_HTTPS_RETURN_URL_NOT_ALLOWED_FOR_A_LIVE_ACCOUNT
          - ONE_TIME_TOKEN_ALREADY_USED
          - ONE_TIME_TOKEN_INVALID
          - RECURRING_CARD_PAYMENTS_NOT_ALLOWED
          - REFUND_AMOUNT_AVAILABLE_MISMATCH
          - REFUND_NOT_AVAILABLE
          - REFUND_NOT_AVAILABLE_DUE_TO_DISPUTE
          - TELEPHONE_PAYMENT_NOTIFICATIONS_NOT_ALLOWED
          - UNEXPECTED_ATTRIBUTE
          - ZERO_AMOUNT_NOT_ALLOWED
          example: GENERIC
        message:
          type: array
          items:
            type: string
            example: error message
        reason:
          type: string
          example: "Optional - ex: amount_not_available"
...
    GatewayAccountRequest:
      type: object
      discriminator:
        propertyName: payment_provider
      properties:
        allow_apple_pay:
          type: boolean
          description: Set to 'true' to enable Apple Pay for this account
          writeOnly: true
        allow_google_pay:
          type: boolean
          description: Set to 'true' to enable Google Pay for this account
          writeOnly: true
        analytics_id:
          type: string
          description: Google Analytics (GA) unique ID for the GOV.UK Pay platform
          example: analytics-id
          writeOnly: true
        description:
          type: string
          description: Some useful non-ambiguous description about the gateway account
          example: account for some gov org
        payment_provider:
          type: string
          default: sandbox
          description: The payment provider for which this account is created
          example: stripe
          writeOnly: true
        requires_3ds:
          type: boolean
          description: Set to 'true' to enable 3DS for this account
          writeOnly: true
        send_payer_email_to_gateway:
          type: boolean
          description: Set to 'true' to enable send payer's email for this account
          writeOnly: true
        send_payer_ip_address_to_gateway:
          type: boolean
          description: Set to 'true' to enable send payer's IP address for this account
          writeOnly: true
        service_id:
          type: string
          example: service-external-id
          writeOnly: true
        service_name:
          type: string
          example: service name
          writeOnly: true
        type:
          type: string
          default: test
          description: Account type for this provider (test/live)
          example: live
          writeOnly: true
...
    GatewayAccountWithCredentialsWithInternalIdResponse:
      type: object
      properties:
        _links:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: string
              format: uri
              example: "{        {            \"href\": \"https://connector.url/v1/api/accounts/1\"\
                ,            \"rel\": \"self\",            \"method\": \"GET\"   \
                \     }    }"
            example: "{        {            \"href\": \"https://connector.url/v1/api/accounts/1\"\
              ,            \"rel\": \"self\",            \"method\": \"GET\"     \
              \   }    }"
          example: "{        {            \"href\": \"https://connector.url/v1/api/accounts/1\"\
            ,            \"rel\": \"self\",            \"method\": \"GET\"       \
            \ }    }"
        allow_apple_pay:
          type: boolean
          default: false
          description: Set to true to enable Apple Pay
          example: true
        allow_authorisation_api:
          type: boolean
          default: false
          description: Flag to indicate whether the account is allowed to initiate
            MOTO payments that are authorised via an API request rather than the web
            interface
          example: true
        allow_google_pay:
          type: boolean
          default: false
          description: Set to true to enable Google Pay
          example: true
        allow_moto:
          type: boolean
          default: false
          description: Indicates whether the Mail Order and Telephone Order (MOTO)
            payments are allowed
        allow_telephone_payment_notifications:
          type: boolean
          default: false
          description: Indicates if the account is used for telephone payments reporting
        allow_zero_amount:
          type: boolean
          default: false
          description: Set to true to support charges with a zero amount
          example: true
        analytics_id:
          type: string
          description: An identifier used to identify the service in Google Analytics.
            The default value is null
        block_prepaid_cards:
          type: boolean
          default: false
          description: Whether pre-paid cards are allowed as a payment method for
            this gateway account
          example: true
        corporate_credit_card_surcharge_amount:
          type: integer
          format: int64
          default: 0
          description: A corporate credit card surcharge amount in pence
          example: 250
        corporate_debit_card_surcharge_amount:
          type: integer
          format: int64
          default: 0
          description: A corporate debit card surcharge amount in pence
          example: 250
        corporate_prepaid_debit_card_surcharge_amount:
          type: integer
          format: int64
          description: A corporate prepaid debit card surcharge amount in pence
          example: 0
        description:
          type: string
          default: "null"
          description: An internal description to identify the gateway account. The
            default value is null.
          example: Account for service xxx
        disabled:
          type: boolean
          default: false
          description: Flag to indicate whether the account is allowed to take payments
            and make refunds
          example: false
        disabled_reason:
          type: string
          description: "The reason the account is disabled, if applicable"
          example: No longer required
        email_collection_mode:
          type: string
          description: "Whether email address is required from paying users. Can be\
            \ MANDATORY, OPTIONAL or OFF"
          enum:
          - MANDATORY
          - OPTIONAL
          - "OFF"
        email_notifications:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/EmailNotificationEntity"
          description: The settings for the different emails (payments/refunds) that
            are sent out
          example:
            PAYMENT_CONFIRMED:
              enabled: true
              template_body: null
              version: 1
            REFUND_ISSUED:
              enabled: true
              template_body: null
              version: 1
        external_id:
          type: string
          description: External ID for the gateway account
          example: fbf905a3f7ea416c8c252410eb45ddbd
        gateway_account_credentials:
          type: array
          description: Array of the credentials configured for this account
          items:
            $ref: "#/components/schemas/GatewayAccountCredentialsWithInternalId"
        gateway_account_id:
          type: integer
          format: int64
          description: The account ID
          example: 1
        integration_version_3ds:
          type: integer
          format: int32
          description: 3DS version used for payments for the gateway account
          example: 2
        live:
          type: boolean
          description: Whether the account is live
          example: true
        moto_mask_card_number_input:
          type: boolean
          default: false
          description: Indicates whether the card number is masked when being input
            for MOTO payments. The default value is false.
        moto_mask_card_security_code_input:
          type: boolean
          default: false
          description: Indicates whether the card security code is masked when being
            input for MOTO payments.
        notifySettings:
          type: object
          additionalProperties:
            type: string
            description: An object containing the Notify credentials and configuration
              for sending custom branded emails
          description: An object containing the Notify credentials and configuration
            for sending custom branded emails
        payment_provider:
          type: string
          description: The payment provider for which this account is created
          example: sandbox
        provider_switch_enabled:
          type: boolean
          default: false
          description: Flag to enable payment provider switching
          example: false
        recurring_enabled:
          type: boolean
          default: false
          description: Flag to indicate whether the account is allowed to take recurring
            card payments
          example: true
        requires3ds:
          type: boolean
          description: Flag to indicate whether 3DS is enabled
          example: true
        send_payer_email_to_gateway:
          type: boolean
          default: false
          description: "If enabled, user email address is included in the authorisation\
            \ request to gateway"
          example: true
        send_payer_ip_address_to_gateway:
          type: boolean
          default: false
          description: "If enabled, user IP address is sent to to gateway"
          example: true
        send_reference_to_gateway:
          type: boolean
          default: false
          description: "If enabled, service payment reference is sent to gateway as\
            \ description. Otherwise payment description is sent to the gateway. Only\
            \ applicable for Worldpay accounts. Default value is 'false'"
          example: true
        service_id:
          type: string
          description: Service external ID
          example: cd1b871207a94a7fa157dee678146acd
        service_name:
          type: string
          description: The service name for the account
          example: service name
        type:
          type: string
          description: Account type for the payment provider (test/live)
          example: test
        worldpay_3ds_flex:
          $ref: "#/components/schemas/Worldpay3dsFlexCredentials"
    GatewayAccountsListDTO:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/GatewayAccountResponse"
...